Cleaned up code and implemented updating radius when zooming
Removed distance normalization. It might be useful in the future since
different screens have different resolutions, but this should work for
now.
Also made the radius now depend on the zoom level. I chose the 100.0 for
the radius and the factor to multiply by arbitrarily, but I think that
they are good for now.