Reorganized code and got lens toggle working
Currently the lens works displaying whatever the second layer is by
"opening" a hole from the first layer. If we want the third layer to be
dispalyed instead of the first, that will require a bit more work (but
should be doable), and that is dependent on the layer system
implementation.