Better error handling for spectral
Modified mipmap_raw_cube and get_raw_data to throw errors when something is wrong when loading the data, instead of the previous approach that returned an empty array.

Modified the usages of get_raw_data to support error handling.

Modified the front and backend of the API to properly communicate the error to the user.