Updated unit tests
Updated get_average_selection test to be updated with new 0-100 scaling.
Discovered implementation was faulty so fixed that.

Updated some testcases in test_routes that test the generation of folders to ensure that the to-be-generated folder is not present before testing. This might occur when a previous run of the test did not remove the folder correctly.