ChangedSpectral element select to periodic table.
Added new file elementSymbols.ts that contains an array with all 118 elements. A seperate file was created because ESLint requires each element to be on a new line, making it annoying to put in other code.

Changed element selection dropdown with a button that shows a popup of the entire periodic table. Periodic table uses IUPAC format.

Changed backend configuration to use new format of symbols (They no longer have 'K' or 'L' as a suffix)

Changed tests to accommodate backend changes.