Added pre-transposing
Added endpoint for pre-transposing fragments such that they can be processed while the user is stiching.
Added transpose_state.py containing a TransposeStateManager responsible for keeping track of the state of each transposed piece.
This ensures that when the stitching request is eventually sent, we can ensure the transposed files are present, or if they failed we can generate them again.